+++ py_trees/display.py
@@ -12,14 +12,6 @@ Behaviour trees are significantly easier to design, monitor and debug
 with visualisations. Py Trees does provide minimal assistance to render
 trees to various simple output formats. Currently this includes dot graphs,
 strings or stdout.
-
-.. warning::
-
-   There is both disrespect for ascii and lack of recognition for unicode
-   in this file as the intention to make ascii art a first class citizen
-   in py_trees became tainted by the desire to make use of the very fine
-   looking unicode symbols underneath. If such behaviour offends, please
-   wear your peril-sensitive sunglasses when parsing or using this module.
 """
 
 ##############################################################################
@@ -41,28 +33,141 @@ from . import utilities
 ##############################################################################
 
 unicode_symbols = {
+    'space': ' ',
+    'bold': console.bold,
+    'bold_reset': console.reset,
     composites.Sequence: u'[-]',
     composites.Selector: u'[o]',
     composites.Parallel: u'[' + console.double_vertical_line + u']',
     decorators.Decorator: u'-^-',
     behaviour.Behaviour: u'-->',
-    common.Status.SUCCESS: console.green + console.check_mark,
-    common.Status.FAILURE: console.red + console.multiplication_x,
-    common.Status.INVALID: console.yellow + u'-',
-    common.Status.RUNNING: console.blue + u'*'
+    common.Status.SUCCESS: console.green + console.check_mark + console.reset,
+    common.Status.FAILURE: console.red + console.multiplication_x + console.reset,
+    common.Status.INVALID: console.yellow + u'-' + console.reset,
+    common.Status.RUNNING: console.blue + u'*' + console.reset
 }
+"""Symbols for a unicode, escape sequence capable console."""
 
 ascii_symbols = {
+    'space': ' ',
+    'bold': console.bold,
+    'bold_reset': console.reset,
     composites.Sequence: "[-]",
     composites.Selector: "[o]",
     composites.Parallel: "[||]",
     decorators.Decorator: "-^-",
     behaviour.Behaviour: "-->",
-    common.Status.SUCCESS: console.green + 'o',
-    common.Status.FAILURE: console.red + 'x',
-    common.Status.INVALID: console.yellow + '-',
-    common.Status.RUNNING: console.blue + '*'
+    common.Status.SUCCESS: console.green + 'o' + console.reset,
+    common.Status.FAILURE: console.red + 'x' + console.reset,
+    common.Status.INVALID: console.yellow + '-' + console.reset,
+    common.Status.RUNNING: console.blue + '*' + console.reset
 }
+"""Symbols for a non-unicode, non-escape sequence capable console."""
+xhtml_symbols = {
+    'space': '<text>&#xa0;</text>',  # &nbsp; is not valid xhtml, see http://www.fileformat.info/info/unicode/char/00a0/index.htm
+    'bold': '<b>',
+    'bold_reset': '</b>',
+    composites.Sequence: '<text>[-]</text>',
+    composites.Selector: '<text>[o]</text>',
+    composites.Parallel: '<text style="color:green;">[&#x2016;]</text>',  # c.f. console.double_vertical_line
+    decorators.Decorator: '<text>-^-</text>',
+    behaviour.Behaviour: '<text>--></text>',
+    common.Status.SUCCESS: '<text style="color:green;">&#x2713;</text>',  # c.f. console.check_mark
+    common.Status.FAILURE: '<text style="color:red;">&#x2715;</text>',  # c.f. console.multiplication_x
+    common.Status.INVALID: '<text style="color:darkgoldenrod;">-</text>',
+    common.Status.RUNNING: '<text style="color:blue;">*</text>'
+}
+"""Symbols for embedding in html."""
+
+
+def _generate_text_tree(
+        root,
+        show_status=False,
+        visited={},
+        previously_visited={},
+        indent=0,
+        symbols=None,
+     ):
+    """
+    Generate a text tree utilising the specified symbol formatter.
+
+    Args:
+        root (:class:`~py_trees.behaviour.Behaviour`): the root of the tree, or subtree you want to show
+        show_status (:obj:`bool`): always show status and feedback message (i.e. for every element,
+            not just those visited)
+        visited (dict): dictionary of (uuid.UUID) and status (:class:`~py_trees.common.Status`) pairs
+            for behaviours visited on the current tick
+        previously_visited (dict): dictionary of behaviour id/status pairs from the previous tree tick
+        indent (:obj:`int`): the number of characters to indent the tree
+        symbols (dict, optional): dictates formatting style
+            (one of :data:`py_trees.display.unicode_symbols` || :data:`py_trees.display.ascii_symbols` || :data:`py_trees.display.xhtml_symbols`),
+            defaults to unicode if stdout supports it, ascii otherwise
+
+    Returns:
+        :obj:`str`: a text-based representation of the behaviour tree
+
+    .. seealso:: :meth:`py_trees.display.ascii_tree`, :meth:`py_trees.display.unicode_tree`, :meth:`py_trees.display.xhtml_tree`
+    """
+    # default to unicode if stdout supports it, ascii otherwise
+    if symbols is None:
+        symbols = unicode_symbols if console.has_unicode() else ascii_symbols
+    tip_id = root.tip().id if root.tip() else None
+
+    def get_behaviour_type(b):
+        for behaviour_type in [composites.Sequence,
+                               composites.Selector,
+                               composites.Parallel,
+                               decorators.Decorator]:
+            if isinstance(b, behaviour_type):
+                return behaviour_type
+        return behaviour.Behaviour
+
+    def style(s, font_weight=False):
+        """
+        Because the way the shell escape sequences reset everything, this needs to get used on any
+        single block of formatted text.
+        """
+        if font_weight:
+            return symbols['bold'] + s + symbols['bold_reset']
+        else:
+            return s
+
+    def generate_lines(root, internal_indent):
+
+        def assemble_single_line(b):
+            font_weight = True if b.id == tip_id else False
+            s = symbols['space'] * 4 * internal_indent
+            s += style(symbols[get_behaviour_type(b)], font_weight)
+            s += " "
+
+            if show_status or b.id in visited.keys():
+                s += style("{} [".format(b.name.replace('\n', ' ')), font_weight)
+                s += style("{}".format(symbols[b.status]), font_weight)
+                message = "" if not b.feedback_message else " -- " + b.feedback_message
+                s += style("]" + message, font_weight)
+            elif (b.id in previously_visited.keys() and
+                  b.id not in visited.keys() and
+                  previously_visited[b.id] == common.Status.RUNNING):
+                s += style("{} [".format(b.name.replace('\n', ' ')), font_weight)
+                s += style("{}".format(symbols[b.status]), font_weight)
+                s += style("]", font_weight)
+            else:
+                s += style("{}".format(b.name.replace('\n', ' ')), font_weight)
+            return s
+
+        if internal_indent == indent:
+            # Root
+            yield assemble_single_line(root)
+            internal_indent += 1
+        for child in root.children:
+            yield assemble_single_line(child)
+            if child.children != []:
+                for line in generate_lines(child, internal_indent + 1):
+                    yield line
+    s = ""
+    for line in generate_lines(root, indent):
+        s += "%s\n" % line
+    return s
 
 
 def ascii_tree(
@@ -73,17 +178,19 @@ def ascii_tree(
         indent=0,
      ):
     """
-    Graffiti your console with with ascii trees.
+    Graffiti your console with ascii art for your trees.
 
     Args:
         root (:class:`~py_trees.behaviour.Behaviour`): the root of the tree, or subtree you want to show
-        indent (:obj:`int`): the number of characters to indent the tree
         show_status (:obj:`bool`): always show status and feedback message (i.e. for every element, not just those visited)
         visited (dict): dictionary of (uuid.UUID) and status (:class:`~py_trees.common.Status`) pairs for behaviours visited on the current tick
         previously_visited (dict): dictionary of behaviour id/status pairs from the previous tree tick
+        indent (:obj:`int`): the number of characters to indent the tree
 
     Returns:
         :obj:`str`: an ascii tree (i.e. in string form)
+
+    .. seealso:: :meth:`py_trees.display.xhtml_tree`, :meth:`py_trees.display.unicode_tree`
 
     Examples:
 
@@ -123,74 +230,109 @@ def ascii_tree(
                                   snapshot_visitor))
             behaviour_tree.visitors.append(snapshot_visitor)
     """
-    symbols = unicode_symbols if console.has_unicode() else ascii_symbols
-    tip_id = root.tip().id if root.tip() else None
-
-    def get_behaviour_type(b):
-        for behaviour_type in [composites.Sequence,
-                               composites.Selector,
-                               composites.Parallel,
-                               decorators.Decorator]:
-            if isinstance(b, behaviour_type):
-                return behaviour_type
-        return behaviour.Behaviour
-
-    def generate_lines(root, internal_indent):
-        if internal_indent == indent:
-            #####################
-            # Root
-            #####################
-            if root.id == tip_id:
-                prefix = "    " * internal_indent + console.bold + symbols[get_behaviour_type(root)] + " "
-            else:
-                prefix = "    " * internal_indent + symbols[get_behaviour_type(root)] + " "
-            if show_status or root.id in visited.keys():
-                message = "" if not root.feedback_message else " -- " + root.feedback_message
-                yield prefix + "{} [{}".format(
-                    root.name.replace('\n', ' '),
-                    symbols[root.status]
-                ) + console.white + "]" + message + console.reset
-            elif (root.id in previously_visited.keys() and
-                  root.id not in visited.keys() and
-                  previously_visited[root.id] == common.Status.RUNNING):
-                yield prefix + root.name.replace('\n', ' ') + " [" + console.yellow + "-" + console.white + "]" + console.reset
-            else:
-                yield prefix + root.name.replace('\n', ' ') + console.reset
-            internal_indent += 1
-        for child in root.children:
-            #####################
-            # Children
-            #####################
-            if child.id == tip_id:
-                prefix = "    " * internal_indent + console.bold + symbols[get_behaviour_type(child)] + " "
-            else:
-                prefix = "    " * internal_indent + symbols[get_behaviour_type(child)] + " "
-            if show_status or child.id in visited.keys():
-                message = "" if not child.feedback_message else " -- " + child.feedback_message
-                yield prefix + child.name.replace('\n', ' ') + " [{}".format(symbols[child.status]) + console.white + "]" + message + console.reset
-            elif (child.id in previously_visited.keys() and
-                  child.id not in visited.keys() and
-                  previously_visited[root.id] == common.Status.RUNNING):
-                yield prefix + child.name.replace('\n', ' ') + " [" + console.yellow + "-" + console.white + "]" + console.reset
-            else:
-                yield prefix + child.name.replace('\n', ' ') + console.reset
-            if child.children != []:
-                for line in generate_lines(child, internal_indent + 1):
-                    yield line
-    s = ""
-    for line in generate_lines(root, indent):
-        s += "%s\n" % line
-    return s
+    lines = _generate_text_tree(
+        root,
+        show_status,
+        visited,
+        previously_visited,
+        indent,
+        symbols=ascii_symbols
+    )
+    return lines
 
 
-def dot_graph(
+def unicode_tree(
+        root,
+        show_status=False,
+        visited={},
+        previously_visited={},
+        indent=0,
+     ):
+    """
+    Graffiti your console with unicode art for your trees.
+
+    Args:
+        root (:class:`~py_trees.behaviour.Behaviour`): the root of the tree, or subtree you want to show
+        show_status (:obj:`bool`): always show status and feedback message (i.e. for every element, not just those visited)
+        visited (dict): dictionary of (uuid.UUID) and status (:class:`~py_trees.common.Status`) pairs for behaviours visited on the current tick
+        previously_visited (dict): dictionary of behaviour id/status pairs from the previous tree tick
+        indent (:obj:`int`): the number of characters to indent the tree
+
+    Returns:
+        :obj:`str`: a unicode tree (i.e. in string form)
+
+    .. seealso:: :meth:`py_trees.display.ascii_tree`, :meth:`py_trees.display.xhtml_tree`
+
+    """
+    lines = _generate_text_tree(
+        root,
+        show_status,
+        visited,
+        previously_visited,
+        indent,
+        symbols=ascii_symbols
+    )
+    return lines
+
+
+def xhtml_tree(
+        root,
+        show_status=False,
+        visited={},
+        previously_visited={},
+        indent=0,
+     ):
+    """
+    Paint your tree on an xhtml snippet.
+
+    Args:
+        root (:class:`~py_trees.behaviour.Behaviour`): the root of the tree, or subtree you want to show
+        show_status (:obj:`bool`): always show status and feedback message (i.e. for every element, not just those visited)
+        visited (dict): dictionary of (uuid.UUID) and status (:class:`~py_trees.common.Status`) pairs for behaviours visited on the current tick
+        previously_visited (dict): dictionary of behaviour id/status pairs from the previous tree tick
+        indent (:obj:`int`): the number of characters to indent the tree
+
+    Returns:
+        :obj:`str`: an ascii tree (i.e. as a xhtml snippet)
+
+    .. seealso:: :meth:`py_trees.display.ascii_tree`, :meth:`py_trees.display.unicode_tree`
+
+    Examples:
+
+    .. code-block:: python
+
+        import py_trees
+        a = py_trees.behaviours.Success()
+        b = py_trees.behaviours.Success()
+        c = c = py_trees.composites.Sequence(children=[a, b])
+        c.tick_once()
+
+        f = open('testies.html', 'w')
+        f.write('<html><head><title>Foo</title><body>')
+        f.write(py_trees.display.xhtml_tree(c, show_status=True))
+        f.write("</body></html>")
+    """
+    lines = _generate_text_tree(
+        root,
+        show_status,
+        visited,
+        previously_visited,
+        indent,
+        symbols=xhtml_symbols
+    )
+    lines = lines.replace("\n", "<br/>\n")
+    return "<code>\n" + lines + "</code>"
+
+
+def dot_tree(
         root: behaviour.Behaviour,
         visibility_level: common.VisibilityLevel=common.VisibilityLevel.DETAIL,
         collapse_decorators: bool=False,
         with_qualified_names: bool=False):
     """
-    Generate the pydot graph - this is usually the first step in
-    rendering the tree to file. See also :py:func:`render_dot_tree`.
+    Paint your tree on a pydot graph.
+
+    .. seealso:: :py:func:`render_dot_tree`.
 
     Args:
         root (:class:`~py_trees.behaviour.Behaviour`): the root of a tree, or subtree
@@ -332,7 +474,7 @@ def render_dot_tree(root: behaviour.Behaviour,
         A good practice is to provide a command line argument for optional rendering of a program so users
         can quickly visualise what tree the program will execute.
     """
-    graph = dot_graph(root, visibility_level, collapse_decorators, with_qualified_names=with_qualified_names)
+    graph = dot_tree(root, visibility_level, collapse_decorators, with_qualified_names=with_qualified_names)
     filename_wo_extension_to_convert = root.name if name is None else name
     filename_wo_extension = utilities.get_valid_filename(filename_wo_extension_to_convert)
     filenames = {}