+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Forthcoming
 **Breaking API**
 
 * [trees] removes the curious looking and unused ``destroy()`` method, `#193 <https://github.com/splintered-reality/py_trees/pull/193>`_
+* [display] ``ascii_tree`` -> ``ascii_tree``/``unicode_tree()``, no longer subverts the choice depending on your stdout, `#192 <https://github.com/splintered-reality/py_trees/pull/192>`_
+* [display] ``dot_graph`` -> ``dot_tree`` for consistency with the text tree methods, `#192 <https://github.com/splintered-reality/py_trees/pull/192>`_
 
 **New Features**