+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Release Notes
 Forthcoming
 -----------
 
+1.2.0 (2019-04-28)
+------------------
+
 **Breaking API**
 
 * [trees] removes the curious looking and unused ``destroy()`` method, `#193 <https://github.com/splintered-reality/py_trees/pull/193>`_
@@ -15,7 +18,7 @@ Forthcoming
 * [behaviour] ``shutdown()`` method to compliment ``setup()``, `#193 <https://github.com/splintered-reality/py_trees/pull/193>`_
 * [display] ``xhtml_tree`` provides an xhtml compatible equivalent to the ``ascii_tree`` representation, `#192 <https://github.com/splintered-reality/py_trees/pull/192>`_
 * [trees] walks the tree calling ``shutdown()`` on each node in it's own ``shutdown()`` method, `#193 <https://github.com/splintered-reality/py_trees/pull/193>`_
-
+* [visitors] get a ``finalise()`` method called immediately prior to post tick handlers, `#191 <https://github.com/splintered-reality/py_trees/pull/191>`_  
 
 1.1.0 (2019-03-19)
 ------------------